refactor(gui): split commands/session.rs into session/{lifecycle,readers,turn}.rs
The 818-line commands/session.rs is split by responsibility into a
directory module with a flat `pub use` re-export façade, so every
external `commands::session::<name>` path (used by the
`generate_handler!` registration in commands/mod.rs and by
voice.rs / voice_android.rs reaching `close_session_inner`) is
unchanged:

  - lifecycle.rs (273) — start / close / resume / list / current-info
    commands + the `close_session_inner` / `prepare_for_session_change`
    / `info_from` teardown+snapshot helpers.
  - readers.rs (338) — the four sidebar reader commands + the pure
    `read_turn_list` / `read_learner` / `read_signals` / `speaker_name`
    / `truncate_to_preview` shape mappers and the display-limit consts.
  - turn.rs (230) — the `send_message` streaming command, the
    `cancel_response` command, `CANCELLED_MESSAGE`, and the
    `run_turn` / `refresh_snapshot` / `cancel_active_turn` helpers.
  - mod.rs (26) — module doc + submodule decls + `pub use *` façade +
    `#[cfg(test)] mod tests;`.

Private helpers the shared test module reaches through `use super::*`
(`run_turn`, `truncate_to_preview`, `info_from`, `refresh_snapshot`,
`cancel_active_turn`, `prepare_for_session_change`,
`RECENT_ENGAGEMENT_DISPLAY_LIMIT`) are bumped from private to
`pub(crate)` so the `pub use` façade re-exports them; the four private
`use` aliases the façade can't carry (`Arc`, `Utc`, `Uuid`, `AppState`)
are imported explicitly in tests.rs. `prepare_for_session_change`'s
`super::voice::…` path becomes `crate::commands::voice::…` now that it
lives one level deeper.

Behaviour-preserving: identical 12-symbol external `pub` surface
(diffed old-vs-new), 207/207 primer-gui tests pass (matches the
pre-split baseline), clippy -D warnings + fmt --check clean.
